Move RateLimiter into separate file
