+++ crates/shared/src/lib.rs
@@ -18,6 +18,7 @@ pub mod network;
 pub mod oneinch_api;
 pub mod paraswap_api;
 pub mod price_estimation;
+pub mod rate_limiter;
 pub mod recent_block_cache;
 pub mod request_sharing;
 pub mod solver_utils;
