+++ crates/shared/src/paraswap_api.rs
@@ -1,12 +1,9 @@
-use crate::{
-    debug_bytes,
-    http_client::{requires_back_off, RateLimiter},
-};
+use crate::{debug_bytes, rate_limiter::RateLimiter};
 use anyhow::Result;
 use derivative::Derivative;
 use ethcontract::{H160, U256};
 use model::u256_decimal;
-use reqwest::{Client, RequestBuilder, Url};
+use reqwest::{Client, RequestBuilder, Response, Url};
 use serde::{
     de::{DeserializeOwned, Error},
     Deserialize, Deserializer, Serialize,
@@ -15,6 +12,11 @@ use serde_json::Value;
 use thiserror::Error;
 
 const BASE_URL: &str = "https://apiv5.paraswap.io";
+
+/// Determines if the HTTP response indicates that the API should back off for a while.
+fn requires_back_off(response: &Result<Response, reqwest::Error>) -> bool {
+    matches!(response, Ok(response) if response.status() == 429)
+}
 
 /// Mockable implementation of the API for unit test
 #[mockall::automock]
